requirements.txt
flask
flask-socketio
pyserial
pystray
Pillow
pyinstaller
pyqt6
platformdirs
requests